[#97] Refactor: Rename .js/.jsx files to .ts/.tsx and update imports - Rename all .js/.jsx files to .ts/.tsx as part of TypeScript migration - Replace relative import paths with alias-based imports for consistency
