+++ src/components/DelayRender.tsx
