+++ src/components/modal/ScriptOntologyModal.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { Alert, Button, Col, Container, Form, Modal, Row } from "react-bootstrap";
-import Rest from "../../rest/Rest.jsx";
+import Rest from "@rest/Rest.tsx";
 
 class ScriptOntologyModal extends React.Component {
   constructor(props) {
