+++ src/components/sform/SFormsModal.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import SForms from "@kbss-cvut/s-forms";
 import { Button, Modal } from "react-bootstrap";
-import Rest from "../../rest/Rest";
-import Loading from "../Loading";
+import Rest from "@rest/Rest.js";
+import Loading from "../Loading.js";
 import "@triply/yasgui/build/yasgui.min.css";
 import ErrorModal from "../modal/ErrorModal";
