+++ src/constants/vocabulary.ts
