+++ src/rest/api.ts
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { API_URL } from "../config/env.js";
+import { API_URL } from "@config/env";
 
 const api = axios.create({
   baseURL: API_URL,
